+++ KVASS/settings.cs
@@ -54,6 +54,17 @@ namespace KVASSNS
         [GameParameters.CustomStringParameterUI("#KVASS_sim_re", lines = 7, title = "#KVASS_sim_re")]
         public string REString { get; private set; } = "";
 
+        public override void SetDifficultyPreset(GameParameters.Preset preset)
+        {
+            
+            switch (preset)
+            {
+                case GameParameters.Preset.Easy:   CareerVessel = 1; ScienceVessel = 0.5f; break;
+                case GameParameters.Preset.Normal: CareerVessel = 2; ScienceVessel = 1; break;
+                case GameParameters.Preset.Hard:   CareerVessel = 5; ScienceVessel = 2;  break;
+            }
+        }
+
         public override bool Enabled(MemberInfo member, GameParameters parameters)
         {
             return true;
@@ -64,16 +75,16 @@ namespace KVASSNS
             if (member == null) return false;
 
             if (member.Name == "IgnoreSPH"
-                || member.Name == "Science_Vessel" || member.Name == "Career_Vessel"
-                || member.Name == "Science_Bureaucracy" || member.Name == "Career_Bureaucracy"
-                || member.Name == "RE_String"
+                || member.Name == "ScienceVessel" || member.Name == "CareerVessel"
+                || member.Name == "ScienceBureaucracy" || member.Name == "CareerBureaucracy"
+                || member.Name == "REString"
                 )
                 return Enable;
 
-            if (member.Name == "Career_Const")
+            if (member.Name == "CareerConst")
                 return CareerBureaucracy && Enable;
 
-            if (member.Name == "Science_Const")
+            if (member.Name == "ScienceConst")
                 return ScienceBureaucracy && Enable;
 
             return true;
@@ -125,17 +136,43 @@ namespace KVASSNS
             minValue = 3, maxValue = 20, stepSize = 1)]
         public int KerbToNextLevel { get; private set; } = 7;
 
+        [GameParameters.CustomParameterUI("#KVASS_plan_enable_sci", toolTip = "#KVASS_plan_enable_sci_tooltip")]
+        public bool SciSpeedUp { get; private set; } = true;
+
+        [GameParameters.CustomIntParameterUI("#KVASS_plan_bureaucracy", minValue = 50, maxValue = 5000, stepSize = 50)]
+        public int SciToNextLevel { get; private set; } = 2500;
+
+
         [GameParameters.CustomParameterUI("#KVASS_plan_enable_bureaucracy", toolTip = "#KVASS_plan_enable_bureaucracy_tooltip")]
         public bool Bureaucracy { get; private set; } = true;
 
         [GameParameters.CustomIntParameterUI("#KVASS_plan_bureaucracy", minValue = 1, maxValue = 142, stepSize = 1)]
         public int BureaucracyTime { get; private set; } = 1;
 
+       
+
         [GameParameters.CustomParameterUI("#KVASS_plan_kill_timewarp", toolTip = "#KVASS_plan_kill_timewarp_tooltip")]
         public bool KillTimeWarp { get; private set; } = false;
 
         [GameParameters.CustomParameterUI("#KVASS_plan_queue", toolTip = "#KVASS_plan_queue_tooltip")]
         public bool Queue { get; private set; } = true;
+
+        [GameParameters.CustomParameterUI("#KVASS_plan_message_speedUps", toolTip = "#KVASS_plan_message_speedUps_tooltip")]
+        public bool ShowMessageSpeedUps { get; private set; } = true;
+
+        public override void SetDifficultyPreset(GameParameters.Preset preset)
+        {
+
+            switch (preset)
+            {
+                case GameParameters.Preset.Easy:   
+                    CareerSeconds = 10; ScienceSeconds = 4; Bureaucracy = false; Queue = false; break;
+                case GameParameters.Preset.Normal: 
+                    CareerSeconds = 10; ScienceSeconds = 4; break;
+                case GameParameters.Preset.Hard:   
+                    CareerSeconds = 20; ScienceSeconds = 8; break;
+            }
+        }
 
         public override bool Enabled(MemberInfo member, GameParameters parameters)
         {
@@ -147,11 +184,12 @@ namespace KVASSNS
             if (member == null) return false;
 
             if (member.Name == "IgnoreSPH"
-                || member.Name == "Career_Seconds" || member.Name == "Science_Seconds"
-                || member.Name == "RepSpeedUp" || member.Name == "KerbSpeedUp"
+                || member.Name == "CareerSeconds" || member.Name == "ScienceSeconds"
+                || member.Name == "RepSpeedUp" || member.Name == "KerbSpeedUp" || member.Name == "SciSpeedUp"
                 || member.Name == "Bureaucracy"
                 || member.Name == "KillTimeWarp"
                 || member.Name == "Queue"
+                || member.Name == "ShowMessageSpeedUps"
 
                 )
                 return Enable;
@@ -161,6 +199,9 @@ namespace KVASSNS
 
             if (member.Name == "KerbToNextLevel")
                 return KerbSpeedUp && Enable;
+
+            if (member.Name == "SciToNextLevel")
+                return SciSpeedUp && Enable;
 
             if (member.Name == "BureaucracyTime")
                 return Bureaucracy && Enable;